Changed type of num_elements to size_t in daos_array writer and reader

+++ benchmarks/daos_array-reader.c
@@ -195,7 +195,7 @@ void read_data(size_t arr_size_mb, int steps, int async) {
   daos_event_t ev, *evp;
   int rc;
   int iter;
-  int num_elements;
+  size_t num_elements;
   daos_size_t size;
   char *eptr;
